Correcting errors in "PagPrincipal.css" and "tacosDePollo.html" style:
In the CSS file, the error detected was that the links were not
centered.On the other hand, in the HTML file a problem was identified
in the application of the font style in the list. This error arose due
to a mismatch in the writing of the style in the CSS file page.
Initially, it was defined as "step01", but in the HTML file, "steps01"
was used. This inconsistency prevented the font style from being
applied correctly. Fortunately, this issue has already been fixed.